requirements.txt
# Dependências necessárias para o backend de Reconhecimento Facial (API)
Flask==3.0.3
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.3.0